Add a few items to outstanding issues in TODO, clean up a lot of warnings (not all, some require more widespread changes than this), and fix a few potential bugs.
git-svn-id: http://anope.svn.sourceforge.net/svnroot/anope/trunk@2457 5417fbe8-f217-4b02-8779-1006273d7864

+++ src/protocol/ratbox.c
@@ -688,8 +688,8 @@ class RatboxProto : public IRCDTS6Proto
 
 		if (!u->nc)
 			return;
-		
-		snprintf(svidbuf, sizeof(svidbuf), "%ld", u->timestamp);
+
+		snprintf(svidbuf, sizeof(svidbuf), "%ld", static_cast<long>(u->timestamp));
 
 		if (u->nc->GetExt("authenticationtoken", c))
 		{
